move default implementation to separate package
